chore(ui): move Tabs component to shared location for reuse across all problems and update import in problem 1
